requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==45.0.4
Django==5.2.3
django-encrypted-model-fields==0.6.5
django-widget-tweaks==1.5.0
Faker==37.4.0
pycparser==2.22
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
